// analytics_dashboard/src/lib/types.ts
export interface OverviewStation {
  bookings_latest_day: number;
  revenue_passenger_paid_latest_day: number;
  revenue_station_fee_latest_day: number;
  real_bookings_latest_day: number;
  ghost_bookings_latest_day: number;
  active_vehicles: number | null;
  trips_latest_day: number | null;
}